distribute OpenCode auth plugin as npm package @onshape-mcp/opencode-auth
Extract the OpenCode auth plugin from .opencode/plugins/ into a proper
npm package so end users can install it via opencode.json config:

  { "plugin": ["@onshape-mcp/opencode-auth"] }

The local dev file becomes a re-export shim that resolves dependencies
from the package's own node_modules/, avoiding dependency duplication.

Package is integrated into the existing release pipeline: version sync,
pack/publish in reflow-release-npm.yml, and staging cleanup.
