+++ npm/opencode-auth/index.ts
@@ -1,0 +1,1 @@
+export { OnshapeAuthPlugin, default } from "./src/plugin.ts";
